GDB Stub: Add monitor commands to manage savestates
GDB supports a "monitor" commands that allows sending target defined
requests to the remote target. The arguments to the "monitor" GDB
command are passed to the stub verbatim.

This adds support for monitor commands to the GDB Stub and implements
monitor commands to load and save states from a numbered slot:

    monitor load <n>
    monitor save <n>