WIP: GDB Stub: Allow detaching and reattaching
This handles the packet GDB sends when detaching and then reinitializes
the GDB stub, waiting for a new connection.

I marked this as WIP, as I'm not confident that this works reliably and
that it correctly reinitializes everything. It also duplicates the
configuration logic from outside GDBStub.cpp. It would likely be better
to refactor the overall initialization.